Chore (cleaning):
moved function to a more appropriate file for a clearer file structure
